requirements.txt
tweepy>=4.10.1
openai
python-dotenv
schedule
requests
beautifulsoup4
python-dateutil
telethon
telethon
python-telegram-bot
python-dotenv
transformers
sentencepiece            # required by MarianMT
torch                   # for the model (CPU or GPU)







